Move stuff into model package
