+++ feature-users/src/main/java/com/jraska/github/client/users/model/GitHubUser.kt
@@ -1,4 +1,4 @@
-package com.jraska.github.client.users
+package com.jraska.github.client.users.model
 
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
