+++ feature-users/src/main/java/com/jraska/github/client/users/model/RepoConverter.kt
@@ -1,21 +1,25 @@
-package com.jraska.github.client.users
+package com.jraska.github.client.users.model
 
 import org.threeten.bp.Instant
 
 internal class RepoConverter {
 
   fun convert(gitHubRepo: GitHubRepo): RepoHeader {
-    return RepoHeader(gitHubRepo.owner!!.login!!,
+    return RepoHeader(
+      gitHubRepo.owner!!.login!!,
       gitHubRepo.name!!, gitHubRepo.description ?: "",
-      gitHubRepo.stargazersCount!!, gitHubRepo.forks!!)
+      gitHubRepo.stargazersCount!!, gitHubRepo.forks!!
+    )
   }
 
   fun convertToDetail(repo: GitHubRepo): RepoDetail {
     val header = convert(repo)
 
     val created = Instant.parse(repo.createdAt)
-    val data = RepoDetail.Data(created, repo.openIssuesCount!!, repo.language!!,
-      repo.subscribersCount!!)
+    val data = RepoDetail.Data(
+      created, repo.openIssuesCount!!, repo.language!!,
+      repo.subscribersCount!!
+    )
 
     return RepoDetail(header, data)
   }
