+++ feature-users/src/main/java/com/jraska/github/client/users/model/RepoDetail.kt
@@ -1,4 +1,4 @@
-package com.jraska.github.client.users
+package com.jraska.github.client.users.model
 
 import org.threeten.bp.Instant
