+++ feature-users/src/main/java/com/jraska/github/client/users/model/User.kt
@@ -1,4 +1,4 @@
-package com.jraska.github.client.users
+package com.jraska.github.client.users.model
 
 class User(
   val login: String,
