+++ feature-users/src/main/java/com/jraska/github/client/users/model/UsersRepository.kt
@@ -1,4 +1,4 @@
-package com.jraska.github.client.users
+package com.jraska.github.client.users.model
 
 import io.reactivex.Observable
 import io.reactivex.Single
