+++ feature-users/src/main/java/com/jraska/github/client/users/ui/RepoDetailHeaderModel.kt
@@ -3,7 +3,7 @@ package com.jraska.github.client.users.ui
 import android.view.View
 import com.airbnb.epoxy.EpoxyModel
 import com.jraska.github.client.users.R
-import com.jraska.github.client.users.RepoDetail
+import com.jraska.github.client.users.model.RepoDetail
 import kotlinx.android.synthetic.main.item_repo_detail_stats.view.repo_detail_created
 import kotlinx.android.synthetic.main.item_repo_detail_stats.view.repo_detail_forks_count
 import kotlinx.android.synthetic.main.item_repo_detail_stats.view.repo_detail_stars_count
