+++ feature-users/src/main/java/com/jraska/github/client/users/ui/UserHeaderModel.kt
@@ -3,7 +3,7 @@ package com.jraska.github.client.users.ui
 import android.view.View
 import com.airbnb.epoxy.EpoxyModel
 import com.jraska.github.client.users.R
-import com.jraska.github.client.users.UserStats
+import com.jraska.github.client.users.model.UserStats
 import kotlinx.android.synthetic.main.item_user_stats.view.user_detail_followers_count
 import kotlinx.android.synthetic.main.item_user_stats.view.user_detail_following_count
 import kotlinx.android.synthetic.main.item_user_stats.view.user_detail_joined
