+++ feature-users/src/test/java/com/jraska/github/client/users/GitHubUserTest.kt
@@ -2,6 +2,7 @@ package com.jraska.github.client.users
 
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
+import com.jraska.github.client.users.model.GitHubUser
 import org.assertj.core.api.Assertions
 import org.junit.Test
 import java.util.ArrayList
