+++ feature-users/src/test/java/com/jraska/github/client/users/RepoDetailViewModelTest.kt
@@ -2,6 +2,8 @@ package com.jraska.github.client.users
 
 import com.jraska.github.client.Navigator
 import com.jraska.github.client.analytics.EventAnalytics
+import com.jraska.github.client.users.model.RepoDetailViewModel
+import com.jraska.github.client.users.model.UsersRepository
 import okhttp3.HttpUrl
 import org.junit.Test
 import org.mockito.Mockito.mock
@@ -12,8 +14,10 @@ class RepoDetailViewModelTest {
   fun whenFullNameClicked_thenLaunchedOnWeb() {
     val navigatorMock = mock(Navigator::class.java)
 
-    val viewModel = RepoDetailViewModel(mock(UsersRepository::class.java),
-      trampoline(), navigatorMock, EventAnalytics.EMPTY)
+    val viewModel = RepoDetailViewModel(
+      mock(UsersRepository::class.java),
+      trampoline(), navigatorMock, EventAnalytics.EMPTY
+    )
 
     viewModel.onFitHubIconClicked("jraska/Falcon")
